moved all code into the Research agent.
